Modify unsubscribe.jsx and nest fetch requests instead of using Promise.all

+++ src/pages/unsubscribe.jsx
@@ -48,43 +48,42 @@ function Unsubscribe() {
   const handleYesClick = (e) => {
     e.preventDefault();
     setUnsubscribing(true);
-    Promise.all([
-      fetch(
-        `https://cauk2n799k.execute-api.eu-west-1.amazonaws.com/dev/api/emails/${id}`,
-        {
-          method: "DELETE",
+    fetch(
+      `https://cauk2n799k.execute-api.eu-west-1.amazonaws.com/dev/api/newsletter/informUnsubscribe`,
+      {
+        method: "POST",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          subscriberId: id,
+        }),
+      }
+    )
+    .then(response => {
+      if (response.ok && response.status === 200) {
+        console.log('email sent')
+        fetch(`https://cauk2n799k.execute-api.eu-west-1.amazonaws.com/dev/api/emails/${id}`, {
+          method: 'DELETE',
           headers: {
             Accept: "application/json",
             "Content-Type": "application/json",
-          },
-        }
-      ),
-      fetch(
-        `https://cauk2n799k.execute-api.eu-west-1.amazonaws.com/dev/api/newsletter/informUnsubscribe`,
-        {
-          method: "POST",
-          headers: {
-            Accept: "application/json",
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            subscriberId: id
-          })
-        }
-      )
-    ])
-    .then(responses => {
-      return Promise.all(responses.map(response => {
-        return response.json();
-      }));
+          }
+        })
+        .then(response => {
+          if (response.ok && response.status === 200) {
+            console.log('booyah deleted from db');
+            setUnsubscribing(false);
+            setFinished(true);
+            return;
+          }
+          throw new Error('Network response was not okay.')
+        })
+      }
     })
-    .then(() => {
-      console.log("booyah");
-      setUnsubscribing(false);
-      setFinished(true);
-    })
-    .catch(error => {
-      console.log(error.message);
+    .catch(err => {
+      console.log(err.message)
       setUnsubscribing(false);
       setError(true);
     })
